github_reminder/webclient/requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.8.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
dateutils==0.6.6
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-import-export==1.2.0
future==0.17.1
gunicorn==19.9.0
idna==2.8
idna-ssl==1.1.0
lxml==4.4.1
nsetools==1.0.10
numpy==1.17.0
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.1.0
pycparser==2.19
PyGithub==1.43.8
Pygments==2.4.2
PyJWT==1.7.1
PyOpenGraph==0.2
python-dateutil==2.8.0
python-telegram-bot==12.0.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.2
telepot==12.7
urllib3==1.25.3
wget==3.2
whitenoise==4.1.3
boto==2.48.0
boto3==1.4.7
botocore==1.10.36
chardet==3.0.4
dj-database-url==0.5.0
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5
django-sslserver==0.20
django-storages==1.5.2
django-widget-tweaks==1.4.1
docutils==0.14
jmespath==0.9.3
Markdown==2.6.11
markdown2==2.3.5
psycopg2==2.7.4
psycopg2-binary==2.7.4
s3transfer==0.1.12